Feat: User, Product Import

+++ buy-gurus-b/src/main/java/com/team04/buy_gurus/orderitem/domain/OrderItem.java
@@ -1,5 +1,7 @@
 package com.team04.buy_gurus.orderitem.domain;
 
+import com.team04.buy_gurus.product.domain.Product;
+import com.team04.buy_gurus.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -16,9 +18,9 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int amount;
+    private Long amount;
 
-    private int price;
+    private Long price;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
@@ -28,7 +30,7 @@ public class OrderItem {
     @JoinColumn(name = "product_id", nullable = false)
     private Product product;
 
-    public OrderItem(int amount, int price, User user, Product product) {
+    public OrderItem(Long amount, Long price, User user, Product product) {
         this.amount = amount;
         this.price = price;
         this.user = user;
